mikutter: fix build against current glib
When compiled against current versions of glib,
the `glib2` gem in `ruby-gnome` fails to build because
of a duplicated declaration of the function
`g_bookmark_file_get_type`.

This has already been fixed upstream [1] [2], but the
respective patch needs a small adjustment to the path
of the affected source file in order to work with
`mikutter`'s gemset declaration.

[1]: https://github.com/ruby-gnome/ruby-gnome/commit/396d2d377bd85d9bdd73b26210e323129deea1f4
[2]: https://github.com/ruby-gnome/ruby-gnome/issues/1547