lib/tests/release.nix: add temporary hook to <nixpkgs/ct>
Ofborg is out-of-repo and has a much smaller (and therefore less
responsive) committer team.

Since it takes a long time to get them to merge and redeploy, this
commit adds a TEMPORARY call from lib/tests/release.nix to the
top-level ct entry point.

This commit should be reverted as soon as ofborg merges and deploys
the following PR, which will cause it to invoke the entry point
directly:

  https://github.com/NixOS/nixpkgs/pull/269356