guile: Fix posix_spawn use on Darwin, Hurd, and other platforms
This backports a patch from current Guile HEAD which fixes a
platform-specific bug causing all subprocess spawns to fail on (at
least) Darwin and GNU Hurd, and possibly other platforms as well.

Upstream patch:
<https://git.savannah.gnu.org/cgit/guix.git/plain/gnu/packages/patches/guile-hurd-posix-spawn.patch>

Note that despite the patch name and comments, other OSes than Hurd
are affected as well.

Thanks to @civodul for the hint that this patch would fix the problem.