nixos/kubo: add Mounts.FuseAllowOther setting for better documentation of the default value
The `Mounts.FuseAllowOther` setting could be used before to override the default but it was not obvious that the default was `true`.

Since this option is irrelevant if the `--mount` flag isn't passed to the daemon (the `autoMount` NixOS option is set to false) the Kubo daemon ignores this setting in this case. This means enabling this option conditionally or not makes no difference in practice, so it might as well be enabled all the time.