sauce-connect: 4.5.4 -> 4.9.1
Latest version on Linux is actually 4.9.2, but latest on Mac is 4.9.1,
so only updating to 4.9.1 for consistency between the two.

Also added passthru.updateScript to keep this package up to date in the
future, with associated refactoring to make that possible.

(cherry picked from commit d31e17fa566a3f5b71933f4e71c025de16219c70)