tmuxPlugins.online-status: unstable-2018-11-30 -> unstable-2018-11-30
Somehow, there was another commit, on the same day as the currently
posted version, that was not tracked in Nixpkgs, in the
tmux-online-status repo. It feels very weird to bump the rev without
changing the version, but I don't know of any derivation revision number
to bump, so I guess this is how it goes.

Signed-off-by: Julia DeMille <me@jdemille.com>