gnome.gpaste: Fix typelib path adjustment
I attempted to fix this in af9e3ddc1da4c96d35b326d20ba09cedb1a5ac3b
but it looks like my mental model of module loader was incorrect so the fix there was insufficient.
Therefore the only option appears to be using a top-level await with a dynamic import.
To avoid having to figure out which imports need to be dynamicized, we just wrap the whole entrypoints.