sparkleshare: Remove fallback to Flatpak image
The upstream .desktop file has a fallback to flatpack. It is only intended to run if `sparkleshare` isn't in the user's path. But to make matters even worse it will also run if `sparkleshare` exists with a failure code.

Remove this fallback to ensure that users don't download and run unpinned and untrusted code from the internet.