maintainer-list: make `github`, `githubId` mandatory fields
Nixpkgs development takes place on 3 platforms:

- GitHub: github.com
- Discourse: discourse.nixos.org
- Matrix: nixos.org server.

As neither Discourse or Matrix are code-oriented platform, we usually expect _at least_
that maintainers are present over GitHub so we can mention them, interact with them via
GitHub automation.

In the future, we may revisit this and remove this restriction if we have
an equivalent silo of data containing information about our own users which we can consume.

In the time being, GitHub remains the only platform against where we can build identity
authentication in the project.