release-python.nix: allow evaluation with openssl_1_1
Home Assistant is a big part of nixpkgs python ecosystem and how I
started participating in python maintenance.

Recently it started including `openssl_1_1` in its transitive dependency
closure, which resulted in its exclusion from python-updates.

This is detrimental to the health of the `home-assistant` ecosystem,
since it would not be built and tested anymore.

Permitting `openssl_1_1` for python-updates, as was done for
`release.nix`, will allow me to get in contact with the upstream, to
hopefully arrive at a favorable outcome.