bash: don't install loadables dir for static build
The dynamic build ships a bunch of sample loadable builtins and installs
them to `$out/lib/bash`. When building statically with musl, the target
fails as it cannot produce any of the dynamically linked loadable
builtins.

Although the failure is not fatal, it also introduces an impurity due to
the heavy parallelization: _sometimes_ the files `loadables.h` and
`Makefile.sample` end up in `$out/lib/bash` depending on how fast the
compilation of the loadable builtins fails. Try the following to
reproduce the issue:

        nix build --rebuild --cores 48 nixpkgs#pkgsStatic.bash

This commit skips building and installing the loadable builtins
altogether if building statically.