buildRustCrate: do not forcibly disable rustc parallelism
buildRustCrate currently sets `codegen-units` to `1` if unspecified.
This option controls the amount of process-level parallelism that
LLVM employs during the build.

The [default is supposed to be
16](https://doc.rust-lang.org/rustc/codegen-options/index.html#codegen-units).
By forcing this to `1` we are disabling all intra-derivation
parallelism.  This makes `buildRustCrate` and `crate2nix` much
slower than `cargo`; we shouldn't do that!