nixos/installer: allow disk image build with `nix.enable=false;`
Building a NixOS disk image results in an error when compiled without Nix tooling,
as indicated by the use of the option `nix.enable = false;`. Disabling Nix tools through
this option not only deactivates general Nix tools but also deletes NixOS installer tools.
Consequently, an error is encountered in the `nixos/lib/make-disk-image.nix` file due to
the absence of the `nixos-install` component.

To address this issue, it is worth noting the availability of the `system.disableInstallerTools`
option, which serves the purpose of disabling installer tools specifically. Therefore, adjusting
the configuration by using `nix.enable = false;` should not interfere with the presence of NixOS
installation tools, as this can be managed separately through `system.disableInstallerTools`.

Signed-off-by: Ganga Ram <Ganga.Ram@tii.ae>