ollama-webui: init at 0.0.0-unstable-2023-12-22
To make open-source large langue models (LLM) accessible there are
projects like Ollama that make it almost trivial to download and
run them locally on a consumer computer. We already have Ollama in
Nixpkgs, but that can only be run conveniently in a terminal
(and doesn't store previous chats). What's missing is a web UI,
e.g. Ollama-WebUI that mimics ChatGPT's frontend and integrates nicely
with Ollama.