grafana: 10.2.2 -> 10.2.3
ChangeLog: https://github.com/grafana/grafana/releases/tag/v10.2.3

Changed `preBuild` to `postConfigure` since the latter is a hook that's
only executed in the actual build, not in the gomod FoD. This is a dirty
hack because apparently the removal of `pkg/util/xorm/go.{mod,sum}`
breaks the module download now, but not removing it breaks the build
later (as documented in the comment above).

A nice side-effect is that we don't do all the `go generate`/`wire`
invocations twice per build.