llama-cpp: assert that only one of the *Support arguments is true
When experimenting with llama-cpp I experienced a mysterious build
error that in the end I tracked down to me setting openclSupport and
openblasSupport to true.  To prevent others from this mistake add an
assert that complains if more than one of the arguments openclSupport,
openblasSupport and rocmSupport is set to true.