fstar: add symlink to ocaml lib
Nix adds the library to OCAMLPATH automatically if the library is
installed in its standard Nix location.

This allows building F* programs extracted to OCaml more easily.
The library has been renamed to `fstar.lib`, though.

Some libraries need to be propagated as they are needed when
building a program that depends on `fstar.lib`.

Fixes #275866