speakeasy-cli: Init at 1.129.1
Source: https://github.com/speakeasy-api/speakeasy
Homepage: https://www.speakeasyapi.dev

This adds the speakeasy-cli tool (speakeasy binary), which
is a tool to manage and create SDKs using the Speakeasy API platform.

Note: that although this is a Golang command, building from source seems
to fail because of private dependencies.