ff2mpv-go: use 'PATH' to invoke 'mpv'
Patching in the full-path to the binary is more robust, unfortunately
this means that a user of `ff2mpv-go` can't use their own wrapped `mpv`
without overriding the package.

As a practical example: using home-manager to configure `mpv` with
additional scripts will *not* work with the package as it was previously
written.

With this commit, we now rely on `PATH` lookups to find the `mpv`
binary, this should prioritize a user's customized `mpv`.