libsoup: Remove passthru.propagatedUserEnvPackages
That does not do anything.

The proper attribute name recognized by the generic builder is `propagatedUserEnvPkgs`
and even that is an ugly hack to have `nix-env` pull in extra packages into the profile.

Looks like it has been mistakenly put to `passthru` since the beginning:
https://github.com/NixOS/nixpkgs/commit/24448ed3de7a2d31ad22922107631ecbdc0fa78c

Apps that use TLS in conjunction with GLib/libsoup should just add `glib-networking` to `buildInputs`
and `wrapGAppsHook` to `nativeBuildInputs` to add the appropriate Gio module
to `GIO_EXTRA_MODULES` environment variable.