transmission_4: remove extraAppArmorPaths
extraAppArmorPaths was added in https://github.com/NixOS/nixpkgs/pull/259663
to extend the default AppArmor ruleset for transmission-daemon. This was
intended to support alternative web interfaces, which is no longer
needed with https://github.com/NixOS/nixpkgs/pull/241247. The NixOS
service was updated to add a user-specified TRANSMISSION_WEB_HOME to the
AppArmor profile without an override.