nixos/yggdrasil: handle /var/lib/yggdrasil on mountpoint
When StateDirectory and DynamicUser are combined, systemd attempts to
move /var/lib/yggdrasil to /var/lib/private/yggdrasil. That does not
work for a mountpoint. At the same time, mounting
/var/lib/private/yggdrasil makes the script that generates
/var/lib/yggdrasil/keys.json do it in the wrong place.

Yggdrasil does not use /var/lib/yggdrasil for any state; therefore it is
unnecessary to pass it as StateDirectory.

To make keys.json accessible to the unit without the StateDirectory
option which causes it to be owned by the correct user, the file is
loaded via LoadCredential.

Finally, the config file preparation steps are improved to actually fail
on errors; as bash does not treat commands that contain $(...) calls
which fail as errors.