chromium: remove maintainer handles from README.md, mention electron
Specifically the maintainers section is quite outdated and prone to get
out of sync with whatever primary data we have (mostly meta.maintainers)
in each derivation.

In an attempt to lower the risk of ending up out of sync again, we
simply remove the maintainer handles.

Also adds a mention for the newly from source built electron variant, as
almost everything except `upstream-info.nix` bumps will trigger electron
rebuilds as well.

And lastly, removes mentions of `chromium{Beta,Dev}` and the
accompanying `google-chrome-{beta,dev}, that have been removed a few
months ago.

I might look into reworking bigger parts of the README.md in the future,
but this honestly isn't that high of a priority for me for now.