linux/hardened: read min supported kernel from kernels-org.json
This one isn't 4.14 anymore and that should've been updated while
removing 4.14, but is easy to miss.

Since it's not expected that we have versions older than the oldest
mainline version from `kernels-org.json`, determine the minimum
supported version by reading it from there.

Also, this means lesser places to update when dropping old kernels.

This needs an additional change for the mainline updater to make sure
that no older versions appear there[1]. This will be implemented in
the next commit.

[1] At the time of implementing this, the oldest supported kernel was
    4.19, however 4.14 wasn't EOL yet and thus still picked up by the
    mainline updater.