tectonic: add passthru.tests.biber-compatibility
The test requires internet access to fetch tectonic's web bundle
on demand. This is achieved by abusing a fixed-output derivation,
which is capable of internet access.

The `tectonic.outPath` is included in the test package name. This
ensures that it is always triggered for rebuild when the main
derivation changes.

Co-authored-by: Doron Behar <doron.behar@gmail.com>