boot.resumeDevice: document this needs to be specified for systemd initrd
According to several sources and my own testing, when the setting
`boot.initrd.systemd.enable = true`, hibernation doesn't work unless
this setting is set. This is even the case when suspending to a swap
device. When using a systemd stage1, the documentation for this option
is misleading as it implies that this should work.