doc/python: update buildPythonApplication example
It is fine to use `with` on the inputs, since that increases the
overall readability of the package.

Removes `wheel` from `nativeBuildInputs`, since it is a result of
cargo culting from an earlier setuptools example, that was wrong, and
it is not required, since it is provided by setuptools itself.