doc: add a jupyter section
we have many kernels available and nixpkgs should be a good way to
configure software as complex as jupyter.
Right now our API is not straightforward and thus needs documenting.
This is not the final doc, it's an initial PR to bootstrap the effort
and refine the nixpkgs API for jupyter.

future work: document how to wrap jupyter-notebook