crosvm: 119.0 -> unstable-2024-01-02
As of a year ago, this project switched to another repository.

I turned on all the features that were listed in [Cargo.toml](https://chromium.googlesource.com/crosvm/crosvm/+/refs/heads/main/Cargo.toml) as working on both aarch64 and x86_64.

This is perhaps too much. No package at the current moment depends on `crosvm`, so maybe it's OK.

I also deleted the updater, as the project no longer appears to have releases, and just deploys out of main. The most recent commit seemed as good as any.