docker: fix LimitNOFILE limit
for the second time, somebody thought setting the NOFILE limit to
`infinity` was a good idea, which breaks a whole lot of systems.

I can’t easily remove the option like the upstream change does, so for
now it sets it to the high-but-bounded number as proposed in the
original PR. We are gonna have to remove once the merged PR lands.